% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msx.R
\name{add_msx}
\alias{add_msx}
\title{Reconstruct original ARIMA model from msx decomposition}
\usage{
add_msx(msx)
}
\arguments{
\item{msx}{An object of class "msx" containing the decomposition of an 
ARIMA model into unobservable component ARIMA models.}
}
\value{
A "um" object representing the reconstructed original ARIMA model.
}
\description{
Reconstructs the original ARIMA model by summing all component models
from an msx decomposition. Used to verify that the decomposition is correct.
}
\examples{
airl <- um(i = "(1-B)(1-B12)")
msx1 <- msx(airl, i = 2)
add_msx(msx1) # returns the airl model

}
\seealso{
\code{\link{add_um}}
}
