% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{AIC.ssm}
\alias{AIC.ssm}
\title{AIC for fitted state space models}
\usage{
\method{AIC}{ssm}(object, k = 2, ...)
}
\arguments{
\item{object}{an object of class \code{\link{ssm}}.}

\item{k}{numeric, penalty per parameter (default is 2).}

\item{...}{additional arguments.}
}
\value{
The AIC value, or \code{NULL} if model is not fitted.
}
\description{
AIC for fitted state space models
}
