% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{altform}
\alias{altform}
\alias{altform.stsm}
\title{Alternative form for STS model}
\usage{
altform(mdl, ...)

\method{altform}{stsm}(mdl, ...)
}
\arguments{
\item{mdl}{an object of class \code{stsm}.}

\item{...}{other arguments.}
}
\value{
An object of class \code{stsm}.
}
\description{
\code{altform} converts a STS model given in contemporaneous form into future
form, and vice versa.
}
\examples{
# Local level model
b <- 1
C <- as.matrix(1)
stsm1 <- stsm(Nile, b, C, S = diag(c(irr = 1, lvl = 0.5)) )
stsm1
stsm2 <- altform(stsm1)

}
