% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ucarima.R
\name{as.ucarima}
\alias{as.ucarima}
\title{Generic function for coercion to class "ucarima"}
\usage{
as.ucarima(object, ...)
}
\arguments{
\item{object}{An object to be coerced.}

\item{...}{Further arguments passed to or from methods.}
}
\value{
An object of class \code{"ucarima"}.
}
\description{
Coerce an object to class "ucarima"
}
