% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{as.um}
\alias{as.um}
\title{Convert \code{arima} into \code{um}.}
\usage{
as.um(arima)
}
\arguments{
\item{arima}{an object of class \code{arima}.}
}
\value{
An object of class \code{um}.
}
\description{
\code{as.um} converts an object of class \code{arima} into an object 
of class \code{um}.
}
\examples{
z <- AirPassengers
a <- arima(log(z), order = c(0,1,1), 
seasonal = list(order = c(0,1,1), frequency = 12))
um1 <- as.um(a)

}
