% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autocov2MA}
\alias{autocov2MA}
\title{MA process compatible with a vector of autocovariances}
\usage{
autocov2MA(x)
}
\arguments{
\item{x}{a vector of q autocovariances.}
}
\value{
A vector of the form c(sig2, 1, -theta1, ..., -thetaq).
}
\description{
\code{autocov2MA} computes the error variance and the MA polynomial from a
vector of autocovriances.
}
\examples{
ma1 <- um(ma = "1 - 0.8B", sig2 = 0.5)
autocov2MA(autocov(ma1, 2))
}
