% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{ccf.tfm}
\alias{ccf.tfm}
\title{Cross-correlation check}
\usage{
ccf.tfm(
  x,
  lag.max = NULL,
  method = c("exact", "cond"),
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{x}{a \code{tfm} object.}

\item{lag.max}{number of lags.}

\item{method}{Exact/conditional residuals.}

\item{envir}{environment in which the function arguments are evaluated.}

\item{...}{additional arguments.}
}
\description{
\code{ccf} displays ccf between prewhitened inputs and residuals.
}
