% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{coef.ucm}
\alias{coef.ucm}
\title{Extract coefficients from UCM objects}
\usage{
\method{coef}{ucm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{ucm}}.}

\item{...}{currently unused.}
}
\value{
A named numeric vector of model coefficients.
}
\description{
Extract coefficients from UCM objects
}
