% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{coef.um}
\alias{coef.um}
\title{Coefficients of a univariate model}
\usage{
\method{coef}{um}(object, ...)
}
\arguments{
\item{object}{a \code{um} object.}

\item{...}{other arguments.}
}
\value{
A numeric vector.
}
\description{
\code{coef} extracts the "coefficients" from a um object.
}
