% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msx.R
\name{.refine_roots}
\alias{.refine_roots}
\title{Refine roots using Newton-Raphson}
\usage{
.refine_roots(roots, coeffs, tol = 1e-15, max_iter = 50)
}
\arguments{
\item{roots}{Inexact roots (real or complex number).}

\item{coeffs}{Numeric vector of polynomial coefficients.}

\item{tol}{Convergence tolerance. Default is 1e-15.}

\item{max_iter}{Maximum number of iterations. Default is 50.}
}
\value{
The refined roots.
}
\description{
\code{.refine_roots} improves the accuracy of the roots (real or complex)
  computed with the \code{polyroot} by using the Newton-Raphson method.
}
\keyword{internal}
