% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucarima.R, R/um.R
\name{equation.ucarima}
\alias{equation.ucarima}
\alias{equation}
\alias{equation.um}
\title{Equation of ucarima model}
\usage{
\method{equation}{ucarima}(x, ...)

equation(x, ...)

\method{equation}{um}(
  x,
  unscramble = FALSE,
  digits = 4,
  z = "z",
  a = "a",
  width = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{um}, a list of these objects or an object
of class \code{ucarima}.}

\item{...}{additional arguments.}

\item{unscramble}{logical. If TRUE, AR, I and MA polynomials are unscrambled.}

\item{digits}{integer. Number of significant digits.}

\item{z}{character. Symbol for time series.}

\item{a}{character. Symbol for error.}

\item{width}{integer. Maximum width for line wrapping. If NULL, uses console width.}
}
\description{
\code{equation} prints the equation of an object of class um.
}
\examples{
equation(um(ar = "(1 - 0.8B)"))

}
