% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{factorize}
\alias{factorize}
\alias{factorize.um}
\title{Factorized form of a univariate ARIMA model}
\usage{
factorize(um, ...)

\method{factorize}{um}(um, full = TRUE, ...)
}
\arguments{
\item{um}{an object of class \code{um} or a list of these objects.}

\item{...}{additional arguments.}

\item{full}{logical value. If TRUE, lag polynomials are completely
factorized. Otherwise, they are factored isolating positive real
roots and grouping the remaining roots.}
}
\description{
\code{factorize} .
}
\examples{
factorize(um(ar = "(1 - 0.8B)"))

}
