% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{fit.stsm}
\alias{fit.stsm}
\title{Estimation of a STS model}
\usage{
\method{fit}{stsm}(mdl, updmdl, par, fixed = NULL, show.iter = FALSE, tol = 1e-04, ...)
}
\arguments{
\item{mdl}{an object of class \code{\link{stsm}}.}

\item{par}{argument passed to the \code{updmod} function.}

\item{fixed}{vector of logical values indicating which parameters are fixed
(TRUE) or estimated (FALSE).}

\item{show.iter}{logical value to show or hide the estimates at the different
iterations.}

\item{tol}{tolerance to check if a root is close to one.}

\item{...}{other arguments.}

\item{updmod}{function to update the parameters of the model.}
}
\value{
An object of class "stsm" with estimated parameters.
}
\description{
\code{fit} is used to estimate a stsm model.
}
\examples{
# Local level model
b <- 1
C <- as.matrix(1)
stsm1 <- stsm(Nile, b, C, s2v = c(lvl = 0.5), s2u = c(irr = 1), fit = FALSE)
stsm1 <- fit(stsm1, method = "L-BFGS-B")
}
