% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{ikf}
\alias{ikf}
\alias{ikf.stsm}
\title{Initialization Kalman filter}
\usage{
ikf(mdl, ...)

\method{ikf}{stsm}(mdl, y = NULL, n = 0, ...)
}
\arguments{
\item{mdl}{an object of class \code{stsm}.}

\item{...}{additional arguments.}

\item{y}{optional time series if it differes from model series.}

\item{n}{integer, number of observations used to estimate the inical
conditions. By default, n is the length of y.}
}
\value{
An list with the initial state and its covariance matrix.
}
\description{
\code{ikf} computes the starting values x0 and P0 by generalized least
squares using the first n observations.
}
