% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{init}
\alias{init}
\alias{init.stsm}
\title{Initial conditions for Kalman filter}
\usage{
init(mdl, ...)

\method{init}{stsm}(mdl, ...)
}
\arguments{
\item{mdl}{an object of class \code{stsm}.}

\item{...}{additional arguments.}
}
\value{
An list with the initial state and its covariance matrix.
}
\description{
\code{init} provides starting values to initialise the Kalman filter
}
