% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{init_kf}
\alias{init_kf}
\title{Initialization of Kalman filter}
\usage{
init_kf(mdl, z = NULL, n = 0)
}
\arguments{
\item{mdl}{an object of class \code{ssm}.}

\item{z}{optional time series if it differs from model series.}

\item{n}{integer, number of observations used to estimate the initial
conditions. If n < d (dimension of state vector), it defaults to length(z).}
}
\value{
A list with components:
\item{x0}{initial state vector estimate}
\item{P0}{covariance matrix of the initial state estimate}
}
\description{
\code{init_kf} computes the starting values x0 and P0 by generalized least
squares using the first n observations.
}
