% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{intervention}
\alias{intervention}
\alias{intervention.tfm}
\title{Intervention analysis/Outlier treatment}
\usage{
intervention(mdl, ...)

\method{intervention}{tfm}(mdl, type, time, n.ahead = 0, envir = NULL, ...)
}
\arguments{
\item{mdl}{an object of class \code{\link{um}} or \code{\link{tfm}}.}

\item{...}{additional arguments.}

\item{type}{the type intervention (pulse, step, ramp) or the type of outlier
(AO, LS, TC, IO).}

\item{time}{the date of the intervention, in format c(year, season).}

\item{n.ahead}{a positive integer to extend the sample period of the
intervention variable with \code{n.ahead} observations, which could be
necessary to forecast the output.}

\item{envir}{environment in which the function arguments are evaluated. If
NULL the parent environment of this function will be used.}
}
\value{
an object of class "\code{\link{tfm}}" or a table.
}
\description{
\code{intervention} estimates the effect of a intervention causing an outlier
at a known time.
}
