% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{inv}
\alias{inv}
\alias{inv.lagpol}
\title{Inverse of a lag polynomial}
\usage{
inv(lp, ...)

\method{inv}{lagpol}(lp, lag.max = 10, ...)
}
\arguments{
\item{lp}{an object of class \code{lagpol}.}

\item{...}{additional arguments.}

\item{lag.max}{largest order of the inverse lag polynomial.}
}
\value{
\code{inv} returns a numeric vector with the coefficients
of the inverse lag polynomial truncated at lag.max.
}
\description{
\code{inv} inverts a lag polynomial until the indicated lag.
}
\examples{
inv(as.lagpol(c(1, 1.2, -0.8))) 

}
