% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{ks}
\alias{ks}
\title{Kalman smoother for SS models}
\usage{
ks(mdl, x0 = NULL, P0 = NULL)
}
\arguments{
\item{mdl}{an object of class \code{ssm}.}

\item{x0}{initial state vector.}

\item{P0}{covariance matrix of x0.}
}
\description{
\code{ks} computes smoothed states and their covariance matrices.
}
