% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{logLik.stsm}
\alias{logLik.stsm}
\title{Log-likelihood of an STS model}
\usage{
\method{logLik}{stsm}(mdl, y = NULL, method = c("exact", "cond"), ...)
}
\arguments{
\item{mdl}{an object of class \code{stsm}.}

\item{y}{an object of class \code{ts}.}

\item{method}{exact or conditional.}

\item{...}{additional parameters.}

\item{tol}{tolerance to check if a root is close to one.}
}
\value{
The exact or conditional log-likelihood.
}
\description{
\code{logLik} computes the exact or conditional log-likelihood of object of 
the class \code{stsm}.
}
