% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{logLik.um}
\alias{logLik.um}
\title{Log-likelihood of an ARIMA model}
\usage{
\method{logLik}{um}(object, z = NULL, method = c("exact", "cond"), ...)
}
\arguments{
\item{object}{an object of class \code{um}.}

\item{z}{an object of class \code{ts}.}

\item{method}{exact or conditional.}

\item{...}{additional arguments.}
}
\value{
The exact or conditional log-likelihood.
}
\description{
\code{logLik} computes the exact or conditional log-likelihood of object of 
the class \code{um}.
}
