% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msx.R
\name{msx}
\alias{msx}
\alias{msx.um}
\title{Minimum signal extraction}
\usage{
msx(mdl, ...)

\method{msx}{um}(
  mdl,
  ar = NULL,
  i = NULL,
  single = FALSE,
  canonical = FALSE,
  cwfact = c("roots", "wilson", "best"),
  pfrac = c("gcd", "solve"),
  tol = 1e-05,
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{mdl}{an object of class \code{\link{um}} or \code{\link{tfm}}.}

\item{...}{additional arguments.}

\item{ar}{autoregressive lag polynomial for the signal.}

\item{i}{integrated lag polynomial for the signal.}

\item{single}{logical. \code{TRUE} for single signal and \code{FALSE} for
multiple signals.}

\item{canonical}{logical value to set or not the canonical requirement.}

\item{cwfact}{method to compute the Cramer-Wold factorization: roots of the
AGCF, Laurie/Wilson algorithms.}

\item{pfrac}{method to compute the partial fraction decomposition of the
spectrum: extended Euclidean algorithm (gcd) or solving linear system (solve).}

\item{tol}{tolerance for null or unit values.}

\item{envir}{the environment.}
}
\value{
An object of class \code{msx}.
}
\description{
\code{msx} extracts a signal from a time series.
}
