% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucarima.R, R/um.R
\name{nabla.ucarima}
\alias{nabla.ucarima}
\alias{nabla}
\alias{nabla.um}
\title{Unscramble I polynomial}
\usage{
\method{nabla}{ucarima}(x, i = NULL, lp = TRUE, ...)

nabla(x, ...)

\method{nabla}{um}(x, ...)
}
\arguments{
\item{x}{an object of class \code{um}.}

\item{i}{integer. Omit this component model.}

\item{lp}{logical indicating the type of return: \code{\link{lagpol}} object
or numeric vector.}

\item{...}{additional arguments.}
}
\value{
A numeric vector \code{c(1, a1, ..., ad)}
}
\description{
\code{nabla} multiplies the I polynomials of an object of 
the \code{um} class.
}
\note{
This function returns the member variable \code{um$nabla}.
}
\examples{
um1 <- um(i = "(1 - B)(1 - B^12)")
nabla(um1)
}
