% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{nabla.stsm}
\alias{nabla.stsm}
\title{Non stationary AR polynomial of the reduced form}
\usage{
\method{nabla}{stsm}(mdl, tol = 1e-04)
}
\arguments{
\item{mdl}{an object of class \code{stsm}.}

\item{tol}{tolerance to check if a root is close to one.}
}
\value{
A numeric vector \code{c(1, a1, ..., ad)}
}
\description{
\code{nabla} returns the non stationary AR polynomialof the reduced form of
 an object of the \code{stsm} class.
}
\examples{
stsm1 <- stsm(b = 1, C = 1, S = diag(c(0.8, 0.04)))
nabla(stsm1)
}
