% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outlierDates}
\alias{outlierDates}
\title{Outlier dates}
\usage{
outlierDates(x, c = 3)
}
\arguments{
\item{x}{an \code{ts} object.}

\item{c}{critical value to determine whether or not an observation is an outlier.}
}
\value{
A table with the indices, dates and z-scores of the outliers.
}
\description{
\code{outlierDates} shows the indeces and dates of outliers.
}
