% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{output}
\alias{output}
\alias{output.tf}
\title{Output of a transfer function or a transfer function model}
\usage{
output(object, ...)

\method{output}{tf}(object, ...)
}
\arguments{
\item{object}{an object of the S3 class "tf".}

\item{...}{additional arguments.}
}
\value{
A "ts" object
}
\description{
\code{output} filters the input using the transfer function.
}
