% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{output.tf}
\alias{output.tf}
\title{Output of a transfer function}
\usage{
output.tf(tf)
}
\arguments{
\item{tf}{an object of the S3 class "tf".}
}
\value{
A "ts" object
}
\description{
\code{output} filters the input using the transfer function.
}
