% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{predict.tf}
\alias{predict.tf}
\title{Predict transfer function}
\usage{
\method{predict}{tf}(object, n.ahead, ...)
}
\arguments{
\item{object}{Transfer function object.}

\item{n.ahead}{Periods to predict.}

\item{...}{Unused.}
}
\value{
Point forecast for input.
}
\description{
Predict transfer function
}
