% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{print.lagpol}
\alias{print.lagpol}
\title{Print Method for Lag Polynomial Objects}
\usage{
\method{print}{lagpol}(x, digits = 2, width = getOption("width"), ...)
}
\arguments{
\item{x}{An object of class \code{lagpol}.}

\item{digits}{Integer. Number of significant digits to display. Default is 2.}

\item{width}{Integer. Maximum number of characters per line. Default is the console width.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the \code{lagpol} object.
}
\description{
Prints objects of class \code{lagpol}.
}
