% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{print.um}
\alias{print.um}
\alias{print.msx}
\title{Print univariate models}
\usage{
\method{print}{um}(x, arima = FALSE, ...)

\method{print}{msx}(x, ...)
}
\arguments{
\item{arima}{logical. If TRUE, lag ARIMA polynomials are printed.}
}
\description{
Print univariate models
}
