% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{print.ssm}
\alias{print.ssm}
\title{Print method for \code{\link{ssm}} objects}
\usage{
\method{print}{ssm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{ssm}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Print method for \code{\link{ssm}} objects
}
