% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{print.summary.ssm}
\alias{print.summary.ssm}
\title{Print summary of fitted state space model}
\usage{
\method{print}{summary.ssm}(x, stats = TRUE, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{summary.ssm}.}

\item{stats}{logical. If \code{TRUE}, additional diagnostic statistics are 
printed.}

\item{digits}{integer. Number of significant digits to display for numeric 
values.}

\item{...}{additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
\code{print.summary.ssm} prints a detailed summary of the results of the
estimation of a \code{\link{ssm}} object.
}
