% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{print.summary.tfm}
\alias{print.summary.tfm}
\title{Print Summary of Transfer Function Model}
\usage{
\method{print}{summary.tfm}(
  x,
  stats = TRUE,
  short = FALSE,
  digits = max(3L, getOption("digits") - 3L),
  ...
)
}
\arguments{
\item{x}{A \code{summary.tfm} object.}

\item{stats}{Logical. If TRUE, prints diagnostic statistics.}

\item{short}{Logical. If TRUE, prints abbreviated output.}

\item{digits}{Number of significant digits.}

\item{...}{Additional arguments.}
}
\description{
Print method for objects of class \code{summary.tfm}.
}
\seealso{
\code{\link{summary.tfm}}
}
