% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{print.tfm}
\alias{print.tfm}
\title{Print Transfer Function Model}
\usage{
\method{print}{tfm}(x, ...)
}
\arguments{
\item{x}{A \code{tfm} object.}

\item{...}{Additional arguments passed to \code{\link{summary.tfm}} 
and \code{\link{print.summary.tfm}}.}
}
\description{
Print method for objects of class \code{tfm}.
}
\details{
Prints a summary of the transfer function model by calling 
\code{summary(x)} with \code{short = TRUE}.
}
\seealso{
\code{\link{tfm}}, \code{\link{summary.tfm}}
}
