% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{print.uc}
\alias{print.uc}
\title{Print method for unobserved components}
\usage{
\method{print}{uc}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{uc}}.}

\item{...}{currently unused.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Print method for unobserved components
}
