% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{printLagpolList}
\alias{printLagpolList}
\title{Prints a list of \code{lagpol} objects.}
\usage{
printLagpolList(llp, digits = 2, width = getOption("width"))
}
\arguments{
\item{llp}{A list of \code{lagpol} objects.}

\item{digits}{Integer. Number of significant digits to display. Default is 2.}

\item{width}{Integer. Maximum number of characters per line. Default is the console width.}
}
\value{
Invisibly returns the list of \code{lagpol} objects.
}
\description{
Prints a list of \code{lagpol} objects.
}
