% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{residuals.tfm}
\alias{residuals.tfm}
\title{Extract Residuals from Transfer Function Model}
\usage{
\method{residuals}{tfm}(
  object,
  y = NULL,
  method = c("exact", "cond"),
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{object}{A fitted \code{tfm} object.}

\item{y}{Optional \code{ts} object for alternative output series.}

\item{method}{Character: "exact" estimates presample values  to
compute residuals; "cond" fixes presample values at zero.}

\item{envir}{Environment for evaluation. NULL uses calling environment.}

\item{...}{Currently unused.}
}
\value{
A \code{ts} object containing model residuals with the same
  time series attributes as the output series.
}
\description{
Computes exact or conditional residuals from a fitted transfer function model.
}
