% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucarima.R
\name{residuals.ucarima}
\alias{residuals.ucarima}
\title{Residuals of fitted UCARIMA models}
\usage{
\method{residuals}{ucarima}(object, method = c("exact", "cond"), ...)
}
\arguments{
\item{object}{an object of class \code{\link{ucarima}}.}

\item{method}{character. Either "exact" or "conditional" residuals.}

\item{...}{additional arguments.}
}
\value{
A \code{ts} object containing the residuals of the SSM.
}
\description{
\code{residuals.ucarima} generates the residuals of a fitted \code{\link{ucarima}}
object.
}
\details{
These residuals are calculated by first converting the
  \code{\link{ucarima}} object to a \code{\link{um}} object and then using the
  \code{\link{residuals.um}} function.
}
