% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{residuals.um}
\alias{residuals.um}
\title{Residuals of the ARIMA model}
\usage{
\method{residuals}{um}(object, z = NULL, method = c("exact", "cond"), ...)
}
\arguments{
\item{object}{an object of class \code{um}.}

\item{z}{an object of class \code{ts}.}

\item{method}{exact/conditional residuals.}

\item{...}{additional arguments.}
}
\value{
An object of class \code{um}.
}
\description{
\code{residuals} computes the exact or conditional residuals.
}
\examples{
z <- AirPassengers
airl <- um(z, i = list(1, c(1, 12)), ma = list(1, c(1, 12)), bc = TRUE)
r <- residuals(airl)
summary(r)
}
