% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{restr.lagpol}
\alias{restr.lagpol}
\title{Restricted lag polynomials}
\usage{
restr.lagpol(param = c(Theta = 0.8), s = 12, f = NULL, p = 1)
}
\arguments{
\item{param}{double, the value of the parameter with a name.}

\item{s}{the seasonal period, integer.}

\item{f}{frequency for second order lag polynomials.}

\item{p}{the power of lag polynomial, integer.}
}
\value{
\code{rest.lagpol} returns an object of class "lagpol".
}
\description{
\code{rest.lagpol} creates two special types of restricted lag polynomials
that can be useful to model seasonal time series:
1 + theta^(1/s)B + theta^(2/s)B^2 + ... + + theta^((s-1)/s)B^{s-1} 
and
1 - 2cos(2pif/s)sqrt(theta)B+ thetaB^2
}
\examples{
rest.lagpol(param = c(Theta = 0.8), s = 12)
rest.lagpol(param = c(Theta = 0.8), s = 12, f = 1)

}
