% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{roots.lagpol}
\alias{roots.lagpol}
\alias{roots.default}
\title{Roots of a lag polynomial}
\usage{
\method{roots}{lagpol}(x, table = TRUE, ...)

\method{roots}{default}(x, ...)
}
\arguments{
\item{x}{an object of class \code{lagpol}.}

\item{table}{logical. If TRUE, it returns a  five columns table showing the
real and imaginary parts, the modulus, the frequency and the period of each
root.}

\item{...}{additional arguments.}
}
\value{
A vector or a table.
}
\description{
\code{roots.lagpol} computes the roots of a lag polynomial.
}
\examples{
roots(c(1, 1.2, -0.8))

}
