% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rsales}
\alias{rsales}
\title{Retail Sales of Variety Stores (U.S. Bureau of the Census)}
\format{
An object of class \code{ts} of length 156.
}
\source{
\url{https://www.census.gov/retail/mrts/mrtshist.html}
}
\usage{
rsales
}
\description{
156 monthly observations from January 1967 to December 1979.
}
\references{
Chen, C. and Liu, L. (1993) Joint Estimation of Model Parameters
and Outlier Effects in Time Series, Journal of the American Statistical
Association, Vol. 88, No. 421, pp. 284-297
}
\keyword{datasets}
