% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{S}
\alias{S}
\title{Annual (rolling) sum}
\usage{
S(x, extend = TRUE)
}
\arguments{
\item{x}{A \code{ts} object.}

\item{extend}{If TRUE, pads result with NAs to match original length.
Default is TRUE.}
}
\value{
Time series of annual sums with same frequency as input.
}
\description{
Computes rolling sum over one year for monthly/quarterly data.
}
