% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detvar.R
\name{sdummies}
\alias{sdummies}
\title{Seasonal dummies}
\usage{
sdummies(Y, ref = 1, constant = FALSE, n.ahead = 0)
}
\arguments{
\item{Y}{A seasonal `ts` object.}

\item{ref}{Reference season (1..frequency).}

\item{constant}{Logical. If `TRUE`, include an intercept column.}

\item{n.ahead}{Integer. Extra observations to extend the sample.}
}
\value{
A `ts` matrix with seasonal dummies (and intercept if requested).
}
\description{
`sdummies()` creates a full set of seasonal dummies (reference-coded).
}
\examples{
D <- sdummies(AirPassengers, ref = 1, constant = TRUE, n.ahead = 24)

}
