% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seriesJ}
\alias{seriesJ}
\title{Gas furnace data}
\format{
A object of class data.frame with 296 rows and 2 columns: 
\describe{
\item{X}{0.60-0.04 (input gas rate in cubir feet per minute.)}
\item{Y}{\% CO2 in outlet gas.}
}
}
\source{
\url{ftp://ftp.wiley.com/public/sci_tech_med/times_series_example}
}
\usage{
seriesJ
}
\description{
Sampling interval 9 seconds; observations for 296 pairs of data points.
}
\references{
Box, G.E., Jenkins, G.M., Reinsel, G.C. and Ljung, G.M. (2015) Time Series
Analysis: Forecasting and Control. John Wiley & Sons, Hoboken.
}
\keyword{datasets}
