% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detvar.R
\name{sincos}
\alias{sincos}
\title{Trigonometric variables}
\usage{
sincos(Y, n.ahead = 0)
}
\arguments{
\item{Y}{an object of class \code{ts} used to determine the sample period and
frequency.}
}
\value{
A matrix of trigonometric variables.
}
\description{
\code{sincos} creates an full set of trigonometric variables.
}
\examples{

Y <- AirPassengers
P58 <- sincos(Y)

}
