% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std}
\alias{std}
\title{Standardize time series}
\usage{
std(x)
}
\arguments{
\item{x}{A \code{ts} object or numeric vector.}
}
\value{
Standardized time series with same class as input.
}
\description{
Centers and scales a time series to zero mean and unit variance.
}
