% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{stsm}
\alias{stsm}
\title{Structural time series models}
\usage{
stsm(y, b, C, S, xreg = NULL, bc = FALSE, cform = TRUE)
}
\arguments{
\item{y}{an object of class \code{ts}.}

\item{b}{vector of constants.}

\item{C}{matrix of constants.}

\item{S}{covariance matrix of the error vector (u_t, v_t).}

\item{xreg}{matrix of regressors.}

\item{bc}{logical. If TRUE logs are taken.}

\item{cform}{logical. If TRUE station equation is given in contemporaneous
form, otherwise it is written in future form.}
}
\value{
An object of class \code{stsm}.
}
\description{
\code{stsm} creates an S3 object representing a time-invariant structural
time series model:
}
\details{
y(t) = b'x(t) + u(t) (observation equation), 
x(t+j) = Cx(t+j-1) + v(t) (state equation),
j = 0 for contemporaneous form or 1 for future form.
}
\examples{
# Local level model
b <- 1
C <- as.matrix(1)
stsm1 <- stsm(Nile, b, C, S = diag(c(irr = 1, lvl = 0.5)) )
stsm1
}
\references{
Durbin, J. and Koopman, S.J. (2012) Time Series Analysis by State Space
Methods, 2nd ed., Oxford University Press, Oxford.

Harvey, A.C. (1989) Forecasting, Structural Time Series Models and the Kalman
Filter. Cambridge University Press, Cambridge.
}
