% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{sum_um}
\alias{sum_um}
\title{Sum of univariate (ARIMA) models}
\usage{
sum_um(...)
}
\arguments{
\item{...}{List of "um" S3 objects.}
}
\value{
A "um" S3 object.
}
\description{
\code{sum_um} creates a univariate (ARIMA) model from the
sum of serveral univariate (arima) models.
}
\examples{
um1 <- um(i = "(1 - B)", ma = "(1 - 0.8B)")
um2 <- um(i = "(1 - B12)", ma = "(1 - 0.8B^12)")
um3 <- sum_um(um1, um2)
}
