% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{summary.ssm}
\alias{summary.ssm}
\title{Summary of fitted state space model}
\usage{
\method{summary}{ssm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{ssm}}.}

\item{...}{additional arguments.}
}
\value{
An object of class \code{summary.ssm}.
}
\description{
\code{summary.ssm} generates a detailed summary of the results of the
estimation of a \code{\link{ssm}} object.
}
