% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{summary.um}
\alias{summary.um}
\title{Summary of um model}
\usage{
\method{summary}{um}(
  object,
  z = NULL,
  method = c("exact", "cond"),
  digits = max(3L, getOption("digits") - 3L),
  ...
)
}
\arguments{
\item{object}{an object of class \code{um}.}

\item{z}{an object of class \code{ts}.}

\item{method}{exact/conditional maximum likelihood.}

\item{digits}{number of significant digits to use when printing.}

\item{...}{additional arguments.}
}
\value{
A list with the summary of the estimation and diagonosis.
}
\description{
\code{summary} prints a summary of the estimation and diagnosis.
}
\examples{
z <- AirPassengers
airl <- um(z, i = list(1, c(1,12)), ma = list(1, c(1,12)), bc = TRUE)
summary(airl)

}
