% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{theta}
\alias{theta}
\alias{theta.um}
\title{Unscramble MA polynomial}
\usage{
theta(um)

\method{theta}{um}(um)
}
\arguments{
\item{um}{an object of class \code{um}.}
}
\value{
A numeric vector \code{c(1, a1, ..., ad)}
}
\description{
Unscramble MA polynomial
}
\note{
This function returns the member variable \code{um$theta}.
}
\examples{
um1 <- um(ma = "(1 - 0.8B)(1 - 0.5B)")
theta(um1)

}
