% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{unitcircle}
\alias{unitcircle}
\alias{unitcircle.default}
\alias{unitcircle.lagpol}
\title{Unit circle}
\usage{
unitcircle(lp, ...)

\method{unitcircle}{default}(x, ...)

\method{unitcircle}{lagpol}(lp, s = 12, ...)
}
\arguments{
\item{lp}{an object of class \code{lagpol}.}

\item{...}{additional arguments.}

\item{s}{integer, seasonal period.}
}
\value{
\code{unitcircle} returns a NULL value.
}
\description{
\code{unitcircle} plots the inverse roots of a lag polynomial together the
unit circle.
}
\examples{
unitcircle(as.lagpol(c(1, rep(0, 11), -1)))

}
