% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wtelephone}
\alias{Wtelephone}
\title{Wisconsin Telephone Company}
\format{
A object of class data.frame with 215 rows and 2 columns: 
\describe{
  \item{X}{Monthly outward station movements.} 
  \item{Y}{Montly inward station movements.} 
}
}
\source{
\url{http://old-www.stat.wisc.edu/sites/default/files/TR222.pdf}
}
\usage{
Wtelephone
}
\description{
Monthly data from January 1951 to October 1966.
}
\references{
Thompson, H. E. and Tiao, G. C. (1971) "Analysis of Telephone
Data: A Case Study of Forecasting Seasonal Time Series," Bell Journal of
Economics, The RAND Corporation, vol. 2(2), pages 515-541, Autumn.
}
\keyword{datasets}
