% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{as.lagpol}
\alias{as.lagpol}
\title{Lag polynomial}
\usage{
as.lagpol(pol, p = 1)
}
\arguments{
\item{pol}{a numeric vector.}

\item{p}{integer power.}
}
\value{
An object of class \code{lagpol}.
}
\description{
\code{as.lagpol} converts a numeric vector c(1, -a_1, \dots, -a_d) into 
a lag polynomial \eqn{(1 - a_1 B - ... - a_p B^p)}.
}
\examples{
as.lagpol(c(1, -0.8))
as.lagpol(c(1, 0, 0, 0, -0.8))
}
