% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{autocov}
\alias{autocov}
\alias{autocov.um}
\title{Theoretical autocovariances of an ARMA model}
\usage{
autocov(um, ...)

\method{autocov}{um}(um, lag.max = 10, ...)
}
\arguments{
\item{um}{an object of class \code{um}.}

\item{...}{additional arguments.}

\item{lag.max}{maximum lag for autocovariances.}
}
\value{
A numeric vector.
}
\description{
\code{autocov} computes the autocovariances of an ARMA model.
}
\note{
The I polynomial is ignored.
}
\examples{
ar1 <- um(ar = "1-0.8B")
autocov(ar1, lag.max = 13)

}
