% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{coef.tfm}
\alias{coef.tfm}
\title{Coefficients of a transfer function model}
\usage{
\method{coef}{tfm}(object, ...)
}
\arguments{
\item{object}{a \code{tfm} object.}

\item{...}{other arguments.}
}
\value{
A numeric vector.
}
\description{
\code{coef} extracts the "coefficients" from a TF model.
}
