% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R, R/tfm.R, R/ucarima.R, R/um.R
\name{decomp.ssm}
\alias{decomp.ssm}
\alias{decomp.tfm}
\alias{decomp.ucarima}
\alias{decomp}
\alias{decomp.um}
\title{Unobserved components}
\usage{
\method{decomp}{ssm}(mdl, tol = 1e-05, ...)

\method{decomp}{tfm}(
  mdl,
  y = NULL,
  method = c("mixed", "forecast", "backcast"),
  envir = NULL,
  ...
)

\method{decomp}{ucarima}(mdl, ...)

decomp(mdl, ...)

\method{decomp}{um}(
  mdl,
  z = NULL,
  method = c("msx", "msx0", "mixed", "forecast", "backcast"),
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{mdl}{an object of class \code{\link{um}} or \code{\link{tfm}}.}

\item{tol}{numeric tolerance for classifying eigenvalues.}

\item{...}{additional arguments.}

\item{y}{an object of class \code{\link{ts}}.}

\item{method}{forward/backward forecasts or a mixture of the two.}

\item{envir}{environment in which the function arguments are evaluated.
If NULL the calling environment of this function will be used.}

\item{z}{an object of class \code{\link{ts}}.}
}
\value{
A matrix with the unobserved components.
}
\description{
\code{decomp} estimates the unobserved components of a time series (trend,
seasonal, cycle, stationary and irregular) from the eventual forecast
function.
}
\examples{
Z <- AirPassengers
um1 <- um(Z, i = list(1, c(1, 12)), ma = list(1, c(1, 12)), bc = TRUE)
#uc1 <- decomp(um1)
}
