% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{equation}
\alias{equation}
\alias{equation.um}
\title{Equation of a univariate ARIMA model}
\usage{
equation(um, ...)

\method{equation}{um}(um, unscramble = TRUE, ...)
}
\arguments{
\item{um}{an object of class \code{um} or a list of these objects.}

\item{...}{additional arguments.}

\item{unscramble}{logical. If TRUE, AR, I and MA polynomials are unscrambled.}
}
\description{
\code{equation} prints the equation of an object of class um.
}
\examples{
equation(um(ar = "(1 - 0.8B)"))

}
