% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{factors}
\alias{factors}
\alias{factors.lagpol}
\title{Lag polynomial factorization}
\usage{
factors(lp, ...)

\method{factors}{lagpol}(lp)
}
\arguments{
\item{lp}{an object of class \code{lagpol}.}
}
\value{
\code{factors} returns a list with the simplifying factors
of the lag polynomial.
}
\description{
\code{factors} extracts the simplifying factors of a polynomial
 in the lag operator.
}
\examples{
factors( as.lagpol(c(1, rep(0, 11), -1)) ) 

}
