% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsm.R
\name{ks}
\alias{ks}
\alias{ks.stsm}
\title{Kalman smoother for STS models}
\usage{
ks(mdl, ...)

\method{ks}{stsm}(mdl, x1 = NULL, P1 = NULL, ...)
}
\arguments{
\item{mdl}{an object of class \code{stsm}.}

\item{...}{additional arguments.}

\item{x1}{initial state vector.}

\item{P1}{covariance matrix of x1.}
}
\value{
An list with the smoothed states and their covariance matrices.
}
\description{
\code{ks} computes smoothed states and their covariance matrices.
}
