% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{msx}
\alias{msx}
\alias{msx.um}
\title{Minimum signal extraction}
\usage{
msx(mdl, ...)

\method{msx}{um}(
  mdl,
  z = NULL,
  ar = NULL,
  i = NULL,
  canonical = TRUE,
  check = TRUE,
  method = c("roots", "acov"),
  ret = c("default", "min"),
  envir = NULL,
  ...
)
}
\arguments{
\item{mdl}{an object of class \code{\link{um}} or \code{\link{tfm}}.}

\item{...}{additional arguments.}

\item{z}{an object of class \code{\link{ts}}.}

\item{ar}{AR polynomial for signal.}

\item{i}{I polynomial for signal.}

\item{canonical}{logical. If TRUE the canonical decomposition is obtained.}

\item{check}{logical value to check if ar and i are simplifying factors of
the object mdl.}

\item{method}{the procedure to obtain the MA parameters of the model of the
signal is based either on the roots or on the autocovariances, method =
c("roots", "acov").}

\item{ret}{type of return, c("default", "min")}

\item{envir}{environment.}
}
\value{
An object of class \code{msx}.
}
\description{
\code{msx} extracts a signal from a time series.
}
