% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{msx}
\alias{msx}
\alias{msx.um}
\title{Minimum signal extraction}
\usage{
msx(mdl, ...)

\method{msx}{um}(
  mdl,
  ar = NULL,
  i = NULL,
  canonical = TRUE,
  tol = 1e-05,
  check = TRUE,
  method = c("roots", "acov"),
  single = TRUE,
  ret = c("default", "min"),
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{mdl}{an object of class \code{\link{um}} or \code{\link{tfm}}.}

\item{...}{additional arguments.}

\item{ar, i}{AR and/or I lag polynomials for the signal.}

\item{canonical}{logical value to set or not the canonical requirement.}

\item{tol}{tolerance to check if a value is null.}

\item{check}{logical value to check if ar and i are simplifying factors of
the object mdl.}

\item{method}{the procedure to obtain the MA parameters of the model of the
signal is based either on the roots or on the autocovariances, method =
c("roots", "acov").}

\item{single}{logical. TRUE for single signal and FALSE for multiple signals.}

\item{ret}{type of return, c("default", "min")}

\item{envir}{environment.}
}
\value{
An object of class \code{msx}.
}
\description{
\code{msx} extracts a signal from a time series.
}
