% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{msx}
\alias{msx}
\alias{msx.um}
\title{Minimum signal extraction}
\usage{
msx(mdl, ...)

\method{msx}{um}(
  mdl,
  z = NULL,
  ar = NULL,
  i = NULL,
  ar.pos = NULL,
  i.pos = NULL,
  canonical = TRUE,
  check = TRUE,
  envir = NULL,
  ...
)
}
\arguments{
\item{mdl}{an object of class \code{\link{um}} or \code{\link{tfm}}.}

\item{...}{additional arguments.}

\item{z}{an object of class \code{\link{ts}}.}

\item{ar}{AR polynomial for signal.}

\item{i}{I polynomial for signal.}

\item{ar.pos}{positions of the AR lag polynomials, vector of integers.}

\item{i.pos}{positions of the I lag polynomials, vector of integers.}

\item{canonical}{logical. If TRUE the canonical decomposition is obtained.}

\item{check}{logical value to check if ar and i are simplifying factors of
the object mdl.}
}
\value{
An object of class \code{msx}.
}
\description{
\code{msx} extracts a signal from a time series.
}
