% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{nabla}
\alias{nabla}
\alias{nabla.um}
\title{Unscramble I polynomial}
\usage{
nabla(um)

\method{nabla}{um}(um)
}
\arguments{
\item{um}{an object of class \code{um}.}
}
\value{
A numeric vector \code{c(1, a1, ..., ad)}
}
\description{
\code{nabla} multiplies the I polynomials of an object of 
the \code{um} class.
}
\note{
This function returns the member variable \code{um$nabla}.
}
\examples{
um1 <- um(i = "(1 - B)(1 - B^12)")
nabla(um1)
}
