% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{noise}
\alias{noise}
\alias{noise.tfm}
\title{Noise of a transfer function model}
\usage{
noise(tfm, ...)

\method{noise}{tfm}(tfm, y = NULL, diff = TRUE, exp = FALSE, ...)
}
\arguments{
\item{tfm}{an object of the class \code{tfm}.}

\item{...}{additional arguments.}

\item{y}{output of the TF model if it is different to that of the 
\code{tfm} object.}

\item{diff}{logical. If TRUE, the noise is differenced with
the "i" operator of the univariate model of the noise.}

\item{exp}{logical. If TRUE, the antilog transformation is applied.}
}
\value{
A "ts" object.
}
\description{
\code{noise} computes the noise of a linear transfer function model.
}
