% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{phi}
\alias{phi}
\alias{phi.um}
\title{Unscramble AR polynomial}
\usage{
phi(um)

\method{phi}{um}(um)
}
\arguments{
\item{um}{an object of class \code{um}.}
}
\value{
A numeric vector \code{c(1, a1, ..., ad)}
}
\description{
\code{phi} multiplies the AR polynomials of an object of 
the \code{um} class.
}
\note{
This function returns the member variable \code{um$phi}.
}
\examples{
um1 <- um(ar = "(1 - 0.8B)(1 - 0.5B)")
phi(um1)

}
