% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{predict.tfm}
\alias{predict.tfm}
\title{Forecasting with transfer function models}
\usage{
\method{predict}{tfm}(object, y = NULL, ori = NULL, n.ahead = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class \code{\link{um}}.}

\item{y}{an object of class \code{\link{ts}}.}

\item{ori}{the origin of prediction. By default, it is the last observation.}

\item{n.ahead}{number of steps ahead.}

\item{level}{confidence level.}

\item{...}{additional arguments.}
}
\description{
\code{predict} computes point and interval predictions for a time series
based on a \code{tfm} object.
}
