% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R, R/ucarima.R, R/um.R
\name{print.tf}
\alias{print.tf}
\alias{print.ucarima}
\alias{print.predict.um}
\alias{print.um}
\title{Print method for transfer function objects}
\usage{
\method{print}{tf}(x, ...)

\method{print}{ucarima}(x, ...)

\method{print}{predict.um}(x, rows = NULL, ...)

\method{print}{um}(x, arima = FALSE, ...)
}
\arguments{
\item{x}{An object of class um.}

\item{...}{Additional arguments.}

\item{rows}{integer. Number of rows printed.}

\item{arima}{logical. If TRUE, lag ARIMA polynomials are printed.}
}
\description{
Print method for transfer function objects

Print ucarima models

Print univariate models

Print univariate models
}
