% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{printLagpol}
\alias{printLagpol}
\title{Print numeric vector as a lagpol object}
\usage{
printLagpol(pol, digits = 2)
}
\arguments{
\item{pol}{numeric vectors with the coefficients of a normalized polynomial.}

\item{digits}{number of decimals.}
}
\description{
Print numeric vector as a lagpol object
}
