% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{printLagpolList}
\alias{printLagpolList}
\title{Print a list of lagpol objects}
\usage{
printLagpolList(llp, digits = 2)
}
\arguments{
\item{llp}{a list of \code{lagpol} objects.}

\item{digits}{number of decimals.}
}
\description{
Print a list of lagpol objects
}
