% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{residuals.tfm}
\alias{residuals.tfm}
\title{Residuals of a transfer function model}
\usage{
\method{residuals}{tfm}(object, y = NULL, method = c("exact", "cond"), ...)
}
\arguments{
\item{object}{a \code{tfm} object.}

\item{y}{output of the TF model (if it is different to that of the "tfm" 
object).}

\item{method}{a character string specifying the method to compute the 
residuals, exact or conditional.}

\item{...}{additional arguments.}
}
\value{
A "ts" object.
}
\description{
\code{residuals} computes the exact or conditional residuals of a TF model.
}
