% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{roots}
\alias{roots}
\alias{roots.um}
\title{Roots of the lag polynomials of an ARIMA model}
\usage{
roots(x, ...)

\method{roots}{um}(x, opr = c("arma", "ar", "ma", "i", "arima"), ...)
}
\arguments{
\item{x}{an object of class \code{um}.}

\item{...}{additional arguments.}

\item{opr}{character that indicates which operators are selected.}
}
\value{
List of matrices with the roots of each single polynomial.
}
\description{
\code{roots} compute the roots of the AR, I, MA lag polynomials an ARIMA
model.
}
\examples{
um1 <- um(ar = "(1 - 0.8B)(1 - 0.8B^12)")
roots(um1)
}
