% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpol.R
\name{roots2lagpol}
\alias{roots2lagpol}
\title{Lag polynomial from roots}
\usage{
roots2lagpol(x, ...)
}
\arguments{
\item{x}{an object of class \code{complex}.}

\item{...}{additional arguments.}
}
\value{
A lag polynomial.
}
\description{
\code{roots2lagpol} creates a lag polynomial from its roots.
}
\examples{
roots2lagpol(polyroot(c(1,-1)))

}
