% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{S}
\alias{S}
\title{Annual sum}
\usage{
S(x, extend = TRUE)
}
\arguments{
\item{x}{an \code{ts} object.}

\item{extend}{logical. If TRUE, the transformed series is extended with NA's
to have the same length as the origianl series.}
}
\value{
The transformed time series, a \code{ts} object.
}
\description{
\code{S} generates the annual sum of a monthly or quarterly time series.
}
