% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detvar.R
\name{sdummies}
\alias{sdummies}
\title{Seasonal dummies}
\usage{
sdummies(Y, ref = 1, n.ahead = 0)
}
\arguments{
\item{Y}{an object of class \code{ts} used to determine the sample period and
frequency.}
}
\value{
A matrix of trigonometric variables.
}
\description{
\code{sincos} creates an full set of seasonal dummies.
}
\examples{

Y <- AirPassengers
P58 <- sincos(Y)

}
