% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detvar.R
\name{sdummies}
\alias{sdummies}
\title{Seasonal dummies}
\usage{
sdummies(Y, ref = 1, constant = FALSE, n.ahead = 0)
}
\arguments{
\item{Y}{an object of class \code{ts} used to determine the sample period and
frequency.}

\item{ref}{the reference season, positive integer}

\item{constant}{logical indicator to include a column of ones.}

\item{n.ahead}{number of additional observations to extend the sample period.}
}
\value{
A matrix of trigonometric variables.
}
\description{
\code{sdummies} creates an full set of seasonal dummies.
}
\examples{

Y <- AirPassengers
P58 <- sincos(Y)

}
