% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{seasadj}
\alias{seasadj}
\alias{seasadj.um}
\title{Seasonal adjustment}
\usage{
seasadj(mdl, ...)

\method{seasadj}{um}(mdl, z = NULL, method = c("mixed", "forecast", "backcast"), ...)
}
\arguments{
\item{mdl}{an object of class \code{\link{um}} or \code{\link{tfm}}.}

\item{...}{additional arguments.}

\item{z}{an object of class \code{\link{ts}}.}

\item{method}{forward/backward forecasts or a mixture of the two.}
}
\value{
\code{seasadj} returns a seasonal adjusted time series.
}
\description{
\code{seasadj} removes the seasonal component of time series.
}
\examples{
Y <- AirPassengers
um1 <- um(Y, bc = TRUE, i = list(1, c(1,12)), ma = list(1, c(1,12)))
Y <- seasadj(um1)
ide(Y)
}
