% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seriesC}
\alias{seriesC}
\title{Series C Chemical Process Temperature Readings: Every Minute.}
\format{
An object of class \code{numeric} of length 226.
}
\source{
\url{ftp://ftp.wiley.com/public/sci_tech_med/times_series_example}
}
\usage{
seriesC
}
\description{
226 observations.
}
\references{
Box, G.E., Jenkins, G.M., Reinsel, G.C. and Ljung, G.M. (2015) Time Series
Analysis: Forecasting and Control. John Wiley & Sons, Hoboken.
}
\keyword{datasets}
