% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{setinputs}
\alias{setinputs}
\alias{setinputs.tfm}
\title{\code{setinputs} adds new inputs into a transfer function model.}
\usage{
setinputs(tfm, ...)

\method{setinputs}{tfm}(tfm, xreg = NULL, inputs = NULL, ...)
}
\arguments{
\item{tfm}{a \code{tfm} object.}

\item{...}{further arguments to be passed to particular methods}

\item{xreg}{a matrix of inputs.}

\item{inputs}{a list of tf objects.}
}
\value{
A \code{tfm} object.
}
\description{
\code{setinputs} adds new inputs into a transfer function model.
}
