% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R, R/um.R
\name{setinputs}
\alias{setinputs}
\alias{setinputs.tfm}
\alias{setinputs.um}
\title{\code{setinputs} adds new inputs into a transfer function model.}
\usage{
setinputs(tfm, ...)

\method{setinputs}{tfm}(mdl, xreg = NULL, inputs = NULL, y = NULL, envir = NULL, ...)

\method{setinputs}{um}(mdl, xreg = NULL, inputs = NULL, y = NULL, envir = NULL, ...)
}
\arguments{
\item{tfm}{a \code{tfm} object.}

\item{...}{other arguments.}

\item{xreg}{a matrix of inputs.}

\item{inputs}{a list of tf objects.}

\item{y}{an optional ts object.}

\item{envir}{an environment.}
}
\value{
A \code{tfm} object.
}
\description{
\code{setinputs} adds new inputs into a transfer function model.
}
