% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R, R/um.R
\name{setinputs.tfm}
\alias{setinputs.tfm}
\alias{setinputs}
\alias{setinputs.um}
\title{\code{setinputs} adds new inputs into a transfer function model.}
\usage{
\method{setinputs}{tfm}(
  mdl,
  xreg = NULL,
  inputs = NULL,
  y = NULL,
  envir = parent.frame(),
  ...
)

setinputs(mdl, ...)

\method{setinputs}{um}(mdl, xreg = NULL, inputs = NULL, y = NULL, envir = NULL, ...)
}
\arguments{
\item{mdl}{a \code{umm} or \code{tfm} object.}

\item{xreg}{a matrix of inputs.}

\item{inputs}{a list of tf objects.}

\item{y}{an optional ts object.}

\item{envir}{an environment.}

\item{...}{other arguments.}
}
\value{
A \code{tfm} object.
}
\description{
\code{setinputs} adds new inputs into a transfer function model.
}
