% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{sform}
\alias{sform}
\alias{sform.um}
\title{Structural form for an ARIMA model}
\usage{
sform(mdl, ...)

\method{sform}{um}(mdl, fSv = NULL, par = NULL, ...)
}
\arguments{
\item{mdl}{an object of class \code{um}.}

\item{...}{other arguments.}

\item{fSv}{optional function to create the covariance matrix.}

\item{par}{vector of parameters for function fSv.}
}
\value{
An object of class \code{stsm}
}
\description{
\code{sform} finds the structural form for an ARIMA model from its the eventual
forecast function.
}
\examples{

airl <- um(i = list(1, c(1, 12)), ma = "(1 - 0.86B)(1 - 0.8B12)")
sf <- sform(airl)
sf

}
