% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{signal}
\alias{signal}
\alias{signal.tfm}
\title{Signal component of a TF model}
\usage{
signal(mdl, ...)

\method{signal}{tfm}(mdl, y = NULL, diff = TRUE, ...)
}
\arguments{
\item{mdl}{an object of the class \code{tfm}.}

\item{...}{additional arguments.}

\item{y}{output of the TF model if it is different to that of the 
\code{tfm} object.}

\item{diff}{logical. If TRUE, the noise is differenced with
the "i" operator of the univariate model of the noise.}
}
\value{
A "ts" object.
}
\description{
\code{signal} extracts the signal of a TF model.
}
