% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R, R/um.R
\name{sim.tfm}
\alias{sim.tfm}
\alias{sim}
\alias{sim.um}
\title{Time series simulation form an ARIMA or TF model}
\usage{
\method{sim}{tfm}(mdl, n = 100, y0 = NULL, seed = NULL, ...)

sim(mdl, ...)

\method{sim}{um}(
  mdl,
  n = 100,
  z0 = NULL,
  n0 = 0,
  a = NULL,
  seed = NULL,
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{mdl}{an object of class \code{um} or \code{tfm}.}

\item{n}{number of observations.}

\item{y0}{initial conditions for the nonstationary series.}

\item{seed}{an integer.}

\item{...}{other arguments.}

\item{z0}{initial conditions for the nonstationary series.}

\item{n0}{remove the n0 first observation, integer.}

\item{a}{vector of innovations, optional.}

\item{envir}{environment in which the function arguments are evaluated.
If NULL the calling environment of this function will be used.}
}
\value{
An object of class \code{ts}.
}
\description{
\code{sim} generates a random time series from an object of class \code{um} or \code{tfm}.
}
