% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detvar.R
\name{sincos}
\alias{sincos}
\title{Trigonometric variables}
\usage{
sincos(Y, n.ahead = 0, constant = FALSE)
}
\arguments{
\item{Y}{A seasonal `ts` object.}

\item{n.ahead}{Integer. Extra observations to extend the sample.}

\item{constant}{Logical. If `TRUE`, include an intercept column.}
}
\value{
A `ts` matrix with trigonometric variables (and intercept if requested).
}
\description{
`sincos()` creates a full set of seasonal trigonometric regressors
(cos/sin harmonics) for a seasonal frequency.
}
\examples{
X <- sincos(AirPassengers, constant = TRUE)

}
