% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{summary.tfm}
\alias{summary.tfm}
\title{Summarizing Transfer Function models}
\usage{
\method{summary}{tfm}(
  object,
  y = NULL,
  method = c("exact", "cond"),
  digits = max(3L, getOption("digits") - 3L),
  envir = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{tfm} object.}

\item{y}{a "ts" object.}

\item{method}{exact or conditional maximum likelihood.}

\item{digits}{number of significant digits to use when printing.}

\item{envir}{environment in which the function arguments are evaluated.
If NULL the calling environment of this function will be used.}

\item{...}{additional arguments.}
}
\value{
A \code{tfm} object.
}
\description{
\code{summary} method for class "tfm".
}
