% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{tfest}
\alias{tfest}
\title{Preestimates of a transfer function}
\usage{
tfest(
  y,
  x,
  delay = 0,
  p = 1,
  q = 2,
  um.y = NULL,
  um.x = NULL,
  n.back = NULL,
  par.prefix = "",
  envir = NULL
)
}
\arguments{
\item{y}{output, a ts object or a numeric vector.}

\item{x}{input, a ts object or a numeric vector.}

\item{delay}{integer.}

\item{p}{order of the AR polynomial, integer}

\item{q}{order of the MA polynomial, integer.}

\item{um.y}{univariate model for output, um object or NULL.}

\item{um.x}{univariate model for input, um object or NULL.}

\item{n.back}{number of backcasts.}

\item{par.prefix}{prefix name for parameters.}

\item{envir}{environment in which the function arguments are evaluated.
If NULL the calling environment of this function will be used.}
}
\value{
A "tf" S3 object
}
\description{
\code{tfest} provides preestimates of the transfer function
between an output and an input.
}
