% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{tfm}
\alias{tfm}
\title{Transfer function models}
\usage{
tfm(output = NULL, xreg = NULL, inputs = NULL, noise, fit = TRUE, ...)
}
\arguments{
\item{output}{a ts object or a numeric vector.}

\item{xreg}{a matrix of regressors.}

\item{inputs}{a list of tf objects.}

\item{noise}{a um object for the noise.}

\item{fit}{logical. If TRUE, model is fitted.}

\item{...}{additional arguments.}
}
\value{
An object of the class \code{tfm}.
}
\description{
\code{tfm} creates a multiple input transfer function model.
}
\references{
Box, G.E., Jenkins, G.M., Reinsel, G.C. and Ljung, G.M. (2015) Time Series
Analysis: Forecasting and Control. John Wiley & Sons, Hoboken.
}
\seealso{
\code{\link{tf}} and \code{\link{um}}.
}
