% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tsvalue}
\alias{tsvalue}
\title{Value of a time series at a date}
\usage{
tsvalue(x, date)
}
\arguments{
\item{x}{an \code{ts} object.}

\item{date}{the time of the specific observation, c(year, month/quarter).}
}
\value{
The value of the observation, double.
}
\description{
\code{tsvalue} select a value from a time series by date.
}
