% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfm.R
\name{varsel}
\alias{varsel}
\alias{varsel.tfm}
\title{Variable selection}
\usage{
varsel(tfm, ...)

\method{varsel}{tfm}(tfm, p.value = 0.1, ...)
}
\arguments{
\item{tfm}{a \code{tfm} object.}

\item{...}{additional arguments.}

\item{p.value}{probability value to decide whether or not to omit an input.}
}
\value{
A \code{tfm} object or a "um" if no input is significant at that level.
}
\description{
\code{varsel} omits non-significant inputs from a transfer function model.
}
