% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um.R
\name{wkfilter}
\alias{wkfilter}
\alias{wkfilter.msx}
\title{Wiener-Kolmogorov filter}
\usage{
wkfilter(x, ...)

\method{wkfilter}{msx}(x, z = NULL, ret = c("signal", "noise"), envir = parent.frame(), ...)
}
\arguments{
\item{x}{an object of class \code{\link{msx}}.}

\item{...}{additional arguments.}

\item{z}{optional, time series.}

\item{ret}{type of return, signal or noise.}

\item{envir}{environment.}
}
\value{
An object of class \code{wkfilter}.
}
\description{
\code{wkfilter} extracts a signal for a time series.
}
